Specify accurate integer data types

+++ cryptokeys.go
@@ -8,19 +8,19 @@ import (
 // Cryptokey structure with JSON API metadata
 type Cryptokey struct {
 	Type       string   `json:"type,omitempty"`
-	ID         uint     `json:"id,omitempty"`
+	ID         uint64   `json:"id,omitempty"`
 	KeyType    string   `json:"keytype,omitempty"`
 	Active     bool     `json:"active,omitempty"`
 	DNSkey     string   `json:"dnskey,omitempty"`
 	DS         []string `json:"ds,omitempty"`
 	Privatekey string   `json:"privatekey,omitempty"`
 	Algorithm  string   `json:"algorithm,omitempty"`
-	Bits       uint     `json:"bits,omitempty"`
+	Bits       uint64   `json:"bits,omitempty"`
 	ZoneHandle *Zone    `json:"-"`
 }
 
-func cryptokeyIDToString(id uint) string {
-	return strconv.FormatUint(uint64(id), 10)
+func cryptokeyIDToString(id uint64) string {
+	return strconv.FormatUint(id, 10)
 }
 
 // GetCryptokeys retrieves a list of Cryptokeys that belong to a Zone
@@ -42,7 +42,7 @@ func (z *Zone) GetCryptokeys() ([]Cryptokey, error) {
 }
 
 // GetCryptokey returns a certain Cryptokey instance of a given Zone
-func (z *Zone) GetCryptokey(id uint) (*Cryptokey, error) {
+func (z *Zone) GetCryptokey(id uint64) (*Cryptokey, error) {
 	cryptokey := new(Cryptokey)
 	myError := new(Error)
 
@@ -57,7 +57,7 @@ func (z *Zone) GetCryptokey(id uint) (*Cryptokey, error) {
 }
 
 // ToggleCryptokey enables/disables a given Cryptokey
-func (z *Zone) ToggleCryptokey(id uint) error {
+func (z *Zone) ToggleCryptokey(id uint64) error {
 	cryptokey := new(Cryptokey)
 	myError := new(Error)
 
@@ -73,7 +73,7 @@ func (c *Cryptokey) ToggleCryptokey() error {
 }
 
 // DeleteCryptokey removes a given Cryptokey
-func (z *Zone) DeleteCryptokey(id uint) error {
+func (z *Zone) DeleteCryptokey(id uint64) error {
 	cryptokey := new(Cryptokey)
 	myError := new(Error)
